Refactor to change mixin syntax
