+++ src/middleware.js
@@ -1,8 +1,7 @@
 'use strict'
 
 const Base = require('./base')
-const Inject = require('./mixins/inject')
 
-module.exports = class Middleware extends Base.mixin(Inject) {
+module.exports = class Middleware extends Base {
 
 }
